local_pipeline/requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
conda==4.3.16
decorator==5.1.1
executing==0.8.3
idna==3.3
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.28
numpy==1.22.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.3
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
stack-data==0.2.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
